backend/requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
pydantic==2.9.2
python-multipart>=0.0.9

qwen-agent[gui]>=0.0.7
dashscope>=1.14.0
pandas>=2.0.0
sqlalchemy>=2.0.0
mysql-connector-python>=8.3.0
python-dotenv>=1.0.0

# 文档助手 —— LlamaIndex 数据层
llama-index-core==0.14.13
llama-index-llms-dashscope>=0.1.0
llama-index-embeddings-dashscope>=0.1.0
llama-index-readers-file>=0.1.0

# 文档助手 —— LangChain 编排层
langchain==1.2.7
langchain-community==0.4.1
langchain-core==1.2.8

# 文档解析
pypdf>=4.0.0
python-docx>=1.1.0
PyMuPDF==1.27.2.2
rapidocr-onnxruntime==1.2.3
Pillow>=10.0.0

# 股票助手
tushare>=1.4.0
pymysql>=1.1.0
